requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
cloudinary==1.29.0
Django==4.1.1
django-cloudinary-storage==0.3.0
django-crispy-forms==1.14.0
django-embed-video==1.4.5
django-tinymce==3.5.0
idna==3.3
Pillow==9.2.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
